model(decoderonly): load embed_tokens in the compile dtype (#674)

+++ src/optimum/rbln/transformers/models/gemma3/modeling_gemma3.py
@@ -405,7 +405,9 @@ class RBLNGemma3ForCausalLM(RBLNDecoderOnlyModelForCausalLM):
                 self.config.pad_token_id,
                 embed_scale=self.config.hidden_size**0.5,
             )
-        return embed_tokens
+        # Gemma3TextScaledWordEmbedding does not forward a dtype kwarg to
+        # nn.Embedding, so cast the module instead.
+        return embed_tokens.to(self.rbln_config.dtype)
 
     @classmethod
     def _update_submodule_config(
